refactor(managed): extract shared ManagedStatusPill
